pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-research-agent-prod"
version = "0.1.0"
description = "Production LangGraph Research Agent - Interactive Test"
requires-python = ">=3.9"

dependencies = [
    "pydantic>=2.0.0",
    "typing-extensions>=4.5.0",
    "python-dotenv>=1.0.0",
    # LangGraph dependencies
    "langgraph>=0.6.0",
    "langchain>=0.3.0", 
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph-checkpoint>=2.0.0",
    # LLM provider
    "langchain-openai>=0.3.0",
    # Research dependencies
    "tavily-python>=0.7.0",
    "wikipedia>=1.4.0",
]

[tool.setuptools.packages.find]
where = ["src"]
